Remove containers and connected components
